Import the 'move' example flow
This is a direct copy of the example flow and documentation present on
docs.globus.org . It comes with some minimal initial structure for
this repo, so that we can begin to experiment with a move of this
content.

Rather than trying to translate the documentation page source itself
into a format suitable for this repo, this duplicates it and converts
to markdown. Only minimal edits are applied to make the text readable
in this repo context.