Add a schema for validating .doc_config.yaml
This is a simple JSON Schema for validating the .doc_config.yaml
files.

Additionally, add a mode for concatenating the `index.adoc` file for
an example page from fragments in the current directory -- this is
anticipated based on examination for the tar-and-transfer example,
which probably needs to be broken down into several files.

Writing the config to allow for this was an initial trial of the
schema's internal consistency and usability. As a side-effect of the
refactor, the 'readme_is_index' config was replaced with a config of
the form

    {"copy": "README.adoc"}

which will probably be the norm for examples, but about which we can
be flexible in the future.

With the schema in place, updates to the doc build script include the
removal of type checks, as we can rely on the schema to do this work.