Build links to the Flows IDE for rendered pages
Update the `build-doc-bundle` script to create a gzip-encoded link to
the example flows in the Flows IDE.

The IDE link is passed to the asciidoc content by setting it as an
attribute of the doc, which the content can then reference in a
`link:` directive.
For now, as a simple matter of simplicity and uniformity, the link is
dropped, unadorned, at the top of each page where it is supported.

Because the compute examples contain multiple definitions and
therefore can't build simple links, an exclude config is supported to
disable the links for these examples.